Better factor namespace module writing + add docs
